Fix negative equality bug RawValueHolder and RichTextValue introduced in 1.2.5.
The problem is that the __ne__ implementation was wrong when the compared objects
were not of the same type.

When __eq__ is called with another type (e.g. None), the implementation
returned NotImplemented.

When __ne__ was called with another type, this lead to not NotImplemented, which is False
althought the result should be true in this case.

The fix is to change __ne__ to also return NotImplemented when the type is different, so
that python can solve the problem.