Fixed deprecation warning for zope.component.interfaces.ComponentLookupError
```
DeprecationWarning: ComponentLookupError is deprecated. Import from zope.interface.interfaces
```

And moved some imports out of the try/except ImportError which is only meant for IMarkupSchema.
With the original code, once the backwards compatibility import in zope.component is gone,
we would have gotten an ImportError for it, and the markup schema would not be available.